backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pypdf
numpy
python-dotenv
openai
google-generativeai
